+++ packages/rocketh/src/index.ts
@@ -10,7 +10,12 @@ export {
 } from './executor/index.js';
 export {createEnvironment, loadDeployments} from './environment/index.js';
 
-export {getChainConfig, chainByCanonicalName, getChainByName} from './environment/utils/chains.js';
+export {
+	getChainConfigFromUserConfigAndDefaultChainInfo,
+	chainByCanonicalName,
+	getDefaultChainInfoByName,
+	getDefaultChainInfoFromChainId,
+} from './environment/utils/chains.js';
 export {mergeArtifacts, mergeABIs} from './environment/utils/artifacts.js';
 export {getGasPriceEstimate, getRoughGasPriceEstimate} from './utils/eth.js';
 export {bigIntToStringReplacer} from './utils/json.js';