+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocketh/doc
 
+## 0.17.1
+
+### Patch Changes
+
+- @rocketh/node@0.17.1
+
 ## 0.17.0
 
 ### Minor Changes