+++ packages/rocketh-router/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/router
 
+## 0.17.1
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.17.1
+  - @rocketh/deploy@0.17.0
+
 ## 0.17.0
 
 ### Minor Changes