+++ packages/rocketh-web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rocketh/web
 
+## 0.17.1
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.17.1
+
 ## 0.17.0
 
 ### Minor Changes